+++ packages/cldr-core/src/index.ts
@@ -12,3 +12,4 @@ export * from './systems';
 export * from './cldr';
 
 export { leftPad } from './utils/string';
+export { checksumIndices } from './resource/checksum';