+++ CHANGELOG.md
@@ -6,6 +6,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [UNRELEASED]
+### Fixed/Changed
+ - Compute a checksum of the schema configuration and add this to the resource packs.
+ - Enforce checksum and version matches at runtime. Any mismatch will throw an error.
+
 ## [0.13.4]
 ### Added
  - New interface for schema customization (experimental). See [example application](https://github.com/phensley/cldr-engine-customization-example)